+++ src/components/HomexDiary/EmotionItem.js
@@ -1,4 +1,4 @@
-import '../../styles/new_DiaryPage/EmotionItem.css'
+import '../../styles/HomexDiary/EmotionItem.css'
 export default function EmotionItem({ id, img, name, onClick, isSelected }){
     const handleOnClick = () => {
         onClick(id);
